src/requirements.txt
elasticsearch==7.8.0
Flask==1.1.2
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
itsdangerous==1.1.0
Flask-Cors==3.0.9
Flask-RESTful==0.3.8
uuid==1.30
selenium==3.141.0
Flask-BasicAuth==0.2.0
Flask-HTTPBasicAuth==1.0.1
pandas==1.2.3
python-dateutil==2.8.1
openpyxl==3.0.6
beautifulsoup4==4.9.3

